set_formatting(), set_field(): Check for null RefPtrs to prevent crashes.
2008-02-03  Murray Cumming  <murrayc@murrayc.com>

* glom/layout_item_dialogs/box_formatting.cc: set_formatting(),
* glom/layout_item_dialogs/dialog_field_layout.cc: set_field():
Check for null RefPtrs to prevent crashes.

* glom/layout_item_dialogs/dialog_field_summary.cc: on_button_field():
Pass the correct parameter to avoid a use of an uninitialized value.
* glom/libglom/data_structure/layout/report_parts/layoutitem_fields_summary.cc:
Constructor: Initialize a member variable to avoid a use of an 
uninitialized value.
This fixed the crashes in bug #513927 (Göran)

svn path=/branches/glom-1-6/; revision=1408

+++ glom/layout_item_dialogs/box_formatting.cc
@@ -108,6 +108,9 @@ void Box_Formatting::set_formatting(const FieldFormatting& format, const Glib::u
   m_combo_choices_field->set_selected_field(choices_field);
   m_combo_choices_field_second->set_selected_field(choices_field_second);
 
+  if(!field)
+    return;
+
   //Custom choices:
   m_adddel_choices_custom->remove_all();
   FieldFormatting::type_list_values list_choice_values = format.get_choices_custom();
